+++ osctld/lib/osctld/container/run_configuration.rb
@@ -17,6 +17,14 @@ module OsCtld
       ctrc
     end
 
+    def self.load_generation(ct, run_id)
+      ctrc = new(ct, load_conf: false, run_id:)
+      return unless File.exist?(ctrc.generation_file_path)
+
+      ctrc.load_conf(path: ctrc.generation_file_path)
+      ctrc
+    end
+
     # @return [Container::RunId]
     attr_reader :run_id
 
@@ -28,15 +36,17 @@ module OsCtld
                                :dist_network_configured
 
     # @param ct [Container]
-    def initialize(ct, load_conf: true)
+    def initialize(ct, load_conf: true, run_id: nil)
       init_lock
       @ct = ct
+      @run_id = run_id
       @cpu_package = nil
       @init_pid = nil
       @aborted = false
       @do_reboot = false
       @exit_promise = Promise.new
       @dist_network_configured = false
+      @generation_cgroups = true
       self.load_conf(from_file: load_conf)
     end
 
@@ -135,19 +145,55 @@ module OsCtld
       File.join('/proc', pid.to_s, 'root')
     end
 
-    attr_writer :aborted
+    def generation_cgroups?
+      inclusively { @generation_cgroups }
+    end
+
+    def cgroup_root
+      generation_cgroups? ? ct.run_cgroup_path(run_id) : ct.base_cgroup_path
+    end
+
+    def cgroup_path
+      generation_cgroups? ? File.join(cgroup_root, 'user-owned') : ct.legacy_cgroup_path
+    end
+
+    def wrapper_cgroup_path
+      generation_cgroups? ? File.join(cgroup_root, 'wrapper') : ct.legacy_wrapper_cgroup_path
+    end
+
+    def lxc_payload_cgroup_path
+      File.join(cgroup_path, generation_cgroups? ? 'payload' : "lxc.payload.#{ct.id}")
+    end
+
+    def lxc_monitor_cgroup_path
+      File.join(cgroup_path, generation_cgroups? ? 'monitor' : "lxc.monitor.#{ct.id}")
+    end
+
+    def lxc_pivot_cgroup_path
+      File.join(cgroup_path, generation_cgroups? ? 'monitor-pivot' : "lxc.pivot.#{ct.id}")
+    end
+
+    def mark_aborted
+      exclusively do
+        @aborted = true
+        save
+      end
+    end
 
     def aborted?
-      @aborted
+      inclusively { @aborted }
     end
 
     # After the current container run stops, start it again
     def request_reboot
-      @do_reboot = true
+      exclusively do
+        @do_reboot = true
+        save
+      end
     end
 
     def reboot?
-      @do_reboot
+      inclusively { @do_reboot }
     end
 
     def get_exit_promise
@@ -172,6 +218,7 @@ module OsCtld
       inclusively do
         {
           'id' => run_id.dump,
+          'generation_cgroups' => generation_cgroups?,
           'dataset' => dataset.to_s,
           'distribution' => distribution,
           'version' => version,
@@ -179,15 +226,18 @@ module OsCtld
           'vendor' => vendor,
           'variant' => variant,
           'cpu_package' => cpu_package,
+          'aborted' => aborted?,
+          'reboot' => reboot?,
           'destroy_dataset_on_stop' => destroy_dataset_on_stop?
         }
       end
     end
 
-    def load_conf(from_file: true)
+    def load_conf(from_file: true, path: nil)
+      source_path = path || file_path
       cfg =
-        if from_file && File.exist?(file_path)
-          OsCtl::Lib::ConfigFile.load_yaml_file(file_path)
+        if from_file && File.exist?(source_path)
+          OsCtl::Lib::ConfigFile.load_yaml_file(source_path)
         else
           {}
         end
@@ -195,8 +245,16 @@ module OsCtld
       @run_id =
         if cfg.has_key?('id')
           Container::RunId.load(cfg['id'])
+        elsif @run_id
+          @run_id
         else
           Container::RunId.new(pool_name: pool.name, container_id: id)
+        end
+      @generation_cgroups =
+        if from_file && File.exist?(source_path)
+          cfg.fetch('generation_cgroups', false)
+        else
+          true
         end
       @dataset =
         if cfg['dataset']
@@ -210,6 +268,8 @@ module OsCtld
       @vendor = cfg['vendor'] || ct.vendor
       @variant = cfg['variant'] || ct.variant
       @cpu_package = cfg['cpu_package']
+      @aborted = cfg.fetch('aborted', false)
+      @do_reboot = cfg.fetch('reboot', false)
       @destroy_dataset_on_stop =
         if cfg.has_key?('destroy_dataset_on_stop')
           cfg['destroy_dataset_on_stop']
@@ -226,15 +286,31 @@ module OsCtld
         # ignore
       end
 
-      regenerate_file(file_path, 0o400) do |new|
-        new.write(OsCtl::Lib::ConfigFile.dump_yaml(dump))
+      dumped = OsCtl::Lib::ConfigFile.dump_yaml(dump)
+
+      [generation_file_path, file_path].each do |path|
+        regenerate_file(path, 0o400) { |new| new.write(dumped) }
       end
     end
 
     def destroy
+      File.unlink(generation_file_path)
+      detach
+    rescue Errno::ENOENT
+      detach
+    end
+
+    # Remove the compatibility pointer only when it still names this run.
+    def detach
+      return unless current_file_matches?
+
       File.unlink(file_path)
     rescue Errno::ENOENT
-      # ignore
+      nil
+    end
+
+    def generation_file_path
+      File.join(dir_path, "config.#{run_id.key}.yml")
     end
 
     protected
@@ -247,6 +323,15 @@ module OsCtld
 
     def file_path
       File.join(dir_path, 'config.yml')
+    end
+
+    def current_file_matches?
+      cfg = OsCtl::Lib::ConfigFile.load_yaml_file(file_path)
+      return false unless cfg['id']
+
+      Container::RunId.load(cfg['id']) == run_id
+    rescue Errno::ENOENT, Psych::Exception, KeyError
+      false
     end
   end
 end
